Clean up tests (move headers out of separate test directory)
